apps/scraper/requirements.txt
# Core Dependencies
playwright==1.40.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Image Processing
Pillow==10.2.0

# Scheduling
apscheduler==3.10.4

# HTTP & Async
httpx==0.26.0
tenacity==8.2.3

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Logging
structlog==24.1.0
